rename bot project setting folder
