+++ Composer/packages/client/src/pages/projectSettings/DeleteBotButton.tsx
